Add AS::Callbacks support to FilterStack.
You can now add before/around/after callbacks to a FilterStack for the :perform and :filter actions. :perform runs when calling ProductFilterStack.perform(scope, params). :filter runs when a single Filter is performed.

There are #scope and #params instance methods on a FilterStack instance so you are able to access and change them in the callbacks.