Rename Exactly to Equal and change its behaviour.
The Exactly Filter filtered on the top level of the params object, as in `{name: 'foo'}`. Now, the default behaviour is to have a `'filter'` key in the params, as in: `{filter: {name: foo}}`.

The key is configurable:

    class ProductFilterStack < Lappen::FilterStack
      use Lappen::Filters::Equal, filter_key: 'filters'
    end

If one want to search on the top level of the params for filterable attributes, provide `nil` as `filter_key`:

    class ProductFilterStack < Lappen::FilterStack
      use Lappen::Filters::Equal, filter_key: nil
    end