Manually extend Rails's autoload_paths when running the FilterStackGenerator if necessary.
When we run the `InstallGenerator`, app/filter_stacks and app/filter_stacks/application_filter_stack.rb are created, but app/filter_stacks is not added to `ActiveSupport::Dependencies.autoload_paths` in the running Spring session.

As the `FilterStackGenerator` checks for existence of `ApplicationFilterStack` in order to subclass from it, Rails's constant autoloading won't find it even if it's there. To circumvent this issue, we check `ActiveSupport::Dependencies.autoload_paths` and add app/filter_stacks if necessary.

If there's any better way to do this, I'm happy to hear from it.